refactor: change code structure and implement config by environment
BREAKING CHANGE:
- reorganizing module base on source and handler
- load config base on environment

revert config modul
