+++ .gitignore
@@ -32,4 +32,4 @@ lerna-debug.log*
 !.vscode/tasks.json
 !.vscode/launch.json
 !.vscode/extensions.json
-.env
+*.env
