+++ src/services/database/city.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { City } from 'src/model/city.entity';
+import { City } from 'src/model/database/city.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
