+++ src/services/database/country.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Country } from 'src/model/country.entity';
+import { Country } from 'src/model/database/country.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
