+++ src/services/database/region.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Region } from 'src/model/region.entity';
+import { Region } from 'src/model/database/region.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
